Convert place to integer, for list index

+++ median.py
@@ -4,9 +4,9 @@ def median(numbers):
     numbers.sort()
     
     if length % 2 == 0:
-        place = length / 2
+        place = int(length / 2) # index must be integer
         calculated_median = (numbers[place] + numbers[place - 1]) / 2.0
     else:
-        place = (length - 1) / 2
+        place = int((length - 1) / 2) # index must be integer
         calculated_median = numbers[place]
     return calculated_median
